함수 위치 이동: ms_find.c 파일로
기존 builtin/unset.c 에 있던 find_prev_node 함수를
->  msfunc/ms_find.c 로 이동하고,

그 내용대로 Makefile, headerfile 수정함
