+++ Makefile
@@ -6,7 +6,7 @@
 #    By: tjung <tjung@student.42.fr>                +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2021/04/22 23:04:25 by minsunki          #+#    #+#              #
-#    Updated: 2022/03/04 03:53:29 by tjung            ###   ########.fr        #
+#    Updated: 2022/03/04 04:09:28 by tjung            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -29,6 +29,7 @@ FIL_M		=	main \
 				msfunc/ms_argjoin \
 				msfunc/ms_isspace \
 				msfunc/ms_correct.c \
+				msfunc/ms_find.c \
 				builtin/pwd \
 				builtin/env \
 				builtin/unset \
